+++ tests/mod.rs
@@ -106,6 +106,10 @@ fn macro_run_basic_tests() {
     let res: ExactNum = expr!(hypot(x, y), &mut ctx);
     debug_assert_eq!(res, x.hypot(&y, p, rm));
 
+    let z = ExactNum::from(2);
+    let res: ExactNum = expr!(fma(x, y, z), &mut ctx);
+    debug_assert_eq!(res, x.fma(&y, &z, p, rm));
+
     let res: ExactNum = expr!(atan2(y, x), &mut ctx);
     debug_assert_eq!(res, y.atan2(&x, p, rm, &mut cc));